move some test files closer to src
