+++ Composer/packages/client/src/utils/convertUtils/__tests__/parsePathToFocused.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License
 
-import { parsePathToFocused } from '../../../src/utils/convertUtils/parsePathToFocused';
+import { parsePathToFocused } from '../parsePathToFocused';
 
 describe('parsePathToFocused', () => {
   it('should return focusedPath', () => {
