+++ Composer/packages/client/src/utils/convertUtils/__tests__/parsePathToSelected.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License
 
-import { parsePathToSelected } from '../../../src/utils/convertUtils/parsePathToSelected';
+import { parsePathToSelected } from '../parsePathToSelected';
 
 describe('parsePathToSelected', () => {
   it('should return selected path', () => {
