+++ Composer/packages/client/src/utils/convertUtils/__tests__/parseTypeToFragment.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License
 import { PromptTab } from '@bfc/shared';
 
-import { parseTypeToFragment } from '../../../src/utils/convertUtils/parseTypeToFragment';
+import { parseTypeToFragment } from '../parseTypeToFragment';
 
 describe('parseTypeToFragment', () => {
   it('should return corrent tab name', () => {
